tests: Moved http interface test to test_http.py
WAZO-4352
