+++ integration_tests/suite/test_config.py
@@ -2,10 +2,9 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import pytest
-import requests
 from hamcrest import assert_that, equal_to, has_entry
 
-from .helpers.base import VALID_TOKEN, APIIntegrationTest
+from .helpers.base import APIIntegrationTest
 
 FAKE_EVENT = {'data': 'Event: foo\r\nAnswerToTheUniverse: 42\r\n\r\n'}
 
@@ -46,39 +45,3 @@ class TestConfigAPI(APIIntegrationTest):
         debug_false_config = self.amid.config()
         assert_that(debug_false_patched_config, equal_to(debug_false_config))
         assert_that(debug_false_config, has_entry('debug', False))
-
-    def _make_raw_http_call(
-        self, verb: str, url: str, body: str | None
-    ) -> requests.Response:
-        headers = {
-            'X-Auth-Token': VALID_TOKEN,
-        }
-
-        match verb:
-            case 'post':
-                call = requests.post
-            case 'patch':
-                call = requests.patch  # type: ignore
-            case _:
-                raise ValueError('Unexpected verb')
-
-        return call(
-            url,
-            headers=headers,
-            data=body,
-            verify=False,
-        )
-
-    def test_that_empty_body_returns_400(self) -> None:
-        port = self.asset_cls.service_port(9491, 'amid')
-        urls = [
-            ('patch', f'http://127.0.0.1:{port}/1.0/config'),
-            ('post', f'http://127.0.0.1:{port}/1.0/action/Command'),
-        ]
-
-        for url in urls:
-            response = self._make_raw_http_call(url[0], url[1], '')
-            assert response.status_code == 400, f'Error with url: {url}'
-
-            response = self._make_raw_http_call(url[0], url[1], None)
-            assert response.status_code == 400, f'Error with url: {url}'
